# lifecat-web/src/main/resources/swaggerapi.yaml
...
swagger: '2.0'
info:
  description: LifeCat项目Version3版本：基于前后端分离，由Springboot + vuejs + WeChat移动端组成，前端通过API接口调用服务。
  version: 1.0.0
  title: LifeCat API
  termsOfService: 'http:///www.lifecat.club/'
  contact:
    email: wshten@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: www.lifecat.club
basePath: /v1
...
schemes:
  - https
  - http
paths:
  /test/all:
    get:
      tags:
        - test
      summary: get all 测试
      description: get all
      operationId: ''
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
          schema:
            $ref: '#/definitions/Pet'
      responses:
        '200':
          description: success
        '400':
          description: post body 为空
        '404':
          description: request 字段缺失
      security:
        - petstore_auth:
            - 'write:pets'
            - 'read:pets'
  /test/list/{testId}:
    get:
      tags:
        - test
      summary: listByUrl 测试
      description: list by id
      operationId: ''
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: testId
          name: testId
          description: no use
          required: true
          schema:
            $ref: '#/definitions/Pet'
      responses:
        '200':
          description: success
        '400':
          description: post body 为空
        '404':
          description: request 字段缺失
      security:
        - petstore_auth:
            - 'write:pets'
            - 'read:pets'
  /test/get/{testId}:
    get:
      tags:
        - test
      summary: get by id 测试
      description: get by id
      operationId: ''
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: testId
          name: testId
          description: test_id
          required: true
          schema:
            $ref: '#/definitions/Pet'
      responses:
        '200':
          description: success
        '400':
          description: post body 为空
        '404':
          description: request 字段缺失
      security:
        - petstore_auth:
            - 'write:pets'
            - 'read:pets'
...
  /test:
    put:
      tags:
        - test
      summary: updateWithEntity test 测试
      description: updateWithEntity test
      operationId: ''
      consumes:
        - application/form
      produces:
        - application/json
      parameters:
        - in: testId
          name: testId
          description: test_id
          required: true
        - in: testName
          name: testName
          description: test_name
          required: true
      responses:
        '200':
          description: success
        '400':
          description: post body 为空
        '404':
          description: request 字段缺失
      security:
        - petstore_auth:
            - 'write:pets'
            - 'read:pets'
  /test/{testId}:
    delete:
      tags:
        - test
      summary: delete test 测试
      description: delete test
      operationId: ''
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: testId
          name: testId
          description: test_id
          required: true
      responses:
        '200':
          description: success
        '400':
          description: post body 为空
        '404':
          description: request 字段缺失
      security:
        - petstore_auth:
            - 'write:pets'
            - 'read:pets'
  /wx/user/login:
    post:
      tags:
        - wechat
      summary: wx用户登录验证
      description: user login
      operationId: ''
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: UserAccountVO Object
          required: true
          schema:
            $ref: '#/definitions/Pet'
      responses:
        '400':
          description: post body 为空
        '404':
          description: request 字段缺失
      security:
        - petstore_auth:
            - 'write:pets'
            - 'read:pets'
  /wx/user:
    post:
      tags:
        - wechat
      summary: Add a new pet to the store
      description: ''
      operationId: addPet
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Pet object that needs to be added to the store
          required: true
          schema:
            $ref: '#/definitions/Pet'
      responses:
        '405':
          description: Invalid input
      security:
        - petstore_auth:
            - 'write:pets'
            - 'read:pets'
    put:
      tags:
        - wechat
      summary: Update an existing pet
      description: ''
      operationId: updatePet
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Pet object that needs to be added to the store
          required: true
          schema:
            $ref: '#/definitions/Pet'
      responses:
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
        '405':
          description: Validation exception
      security:
        - petstore_auth:
            - 'write:pets'
            - 'read:pets'
...
  /user/login:
    get:
      tags:
        - user
      summary: Logs user into the system
      description: ''
      operationId: loginUser
      produces:
        - application/xml
        - application/json
      parameters:
        - name: username
          in: query
          description: The user name for login
          required: true
          type: string
        - name: password
          in: query
          description: The password for login in clear text
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: string
          headers:
            X-Rate-Limit:
              type: integer
              format: int32
              description: calls per hour allowed by the user
            X-Expires-After:
              type: string
              format: date-time
              description: date in UTC when token expires
        '400':
          description: Invalid username/password supplied
...
  '/user/{username}':
    get:
      tags:
        - user
      summary: Get user by user name
      description: ''
      operationId: getUserByName
      produces:
        - application/xml
        - application/json
      parameters:
        - name: username
          in: path
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
    put:
      tags:
        - user
      summary: Updated user
      description: This can only be done by the logged in user.
      operationId: updateUser
      produces:
        - application/xml
        - application/json
      parameters:
        - name: username
          in: path
          description: name that need to be updated
          required: true
          type: string
        - in: body
          name: body
          description: Updated user object
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '400':
          description: Invalid user supplied
        '404':
          description: User not found
    delete:
      tags:
        - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      produces:
        - application/xml
        - application/json
      parameters:
        - name: username
          in: path
          description: The name that needs to be deleted
          required: true
          type: string
      responses:
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
securityDefinitions:
  petstore_auth:
    type: oauth2
    authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
    flow: implicit
    scopes:
      'write:pets': modify pets in your account
      'read:pets': read your pets
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
...
  Category:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
    xml:
      name: Category
  User:
    type: object
    properties:
      id:
        type: integer
        format: int64
      username:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      userStatus:
        type: integer
        format: int32
        description: User Status
    xml:
      name: User
  Tag:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
    xml:
      name: Tag
  Pet:
    type: object
    required:
      - name
      - photoUrls
    properties:
      id:
        type: integer
        format: int64
      category:
        $ref: '#/definitions/Category'
      name:
        type: string
        example: doggie
      photoUrls:
        type: array
        xml:
          name: photoUrl
          wrapped: true
        items:
          type: string
      tags:
        type: array
        xml:
          name: tag
          wrapped: true
        items:
          $ref: '#/definitions/Tag'
      status:
        type: string
        description: pet status in the store
        enum:
          - available
          - pending
          - sold
    xml:
      name: Pet
...